feat: A Bluesky feed for norwegian languages
Implements an ATProto feed server that subscribes to the Bluesky
firehose and keeps track of all posts written in languages used in
Norway. This includes norwegian bokmål, nynorsk, and sami languages.

The server is implemented as a golang app that exposes commands for
doing various actions. This includes serve, which serves the feed server
and subscribes to the firehose, the publish command which publishes the
feed to the Bluesky server, etc.

Norsky is architected as several go routines that subscribe or publish
to go channels. For example the firehose will be initialized with a go
channel for passing event structs that tells the app that a post has
been created or deleted. A db writer go routine subscribes to the
channel and writes the post to or deletes it from the db.

SQLite is used as a db because it is simple and does not require much
setup. This should suffice for the norwegian language community of bsky
as there are not that many posts.

Goreleaser is used to create releases. Binaries are built for arm64,
amd64, and linux and windows. It also builds docker images to make it
simple to run Norsky in a docker container.

Currently there are no tests because this is a fun hobby project and not
a serious project with paying customers. Tests might be added later.

Implement initial server handlers for atproto feed specs.

describeFeedGenerator list the did of the feed generator and
provide a list of feeds that can be subscribed to.

getFeedSkeleton executes the feed logic itself getting the feed to
execute from the query parameter.

The server and algorithms are still work in progress, but getting
closer.

Use viper to provide environment variable and config file support. Cobra
and Viper does not work 100% well together, but it is close enough that
parameters can be provided and picked up in the commands relatively
easily.

Adds migration command to run db migrations as a one off task. Useful
when testing application.

Add Favicon and image to use as a feed avatar.

Todo: Add publish functionality, complete the server part, and figure
out the remaining atproto/bluesky feed spec stuff. Lots to learn.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>

Unfinished publish command and better feed endpoint

The feed endpoint not returns the feed skeleton and handles cursors and
all very well. Changed SQL to aggregate up languages to make limit and
cursors work correctly.

Cursor parsing now happens inside algorithms package so the fiber server
logic can be completely ignorant of how cursors should work. This seems
like more of a bluesky algorithm than server thing.

Realized Bluesky has no, nb, and nn language codes meaning some users
might not define if they write in bokmål or nynorsk. For now the feed
will only return no posts for the algorithm that returns all norwegian
languages.

Todo:

- Switch cli stuff to urfave cli to simplify command setup and env
  fallback
- Finish setting up feed generator publisher based on the work by
  Furrysky feed generator server  https://github.com/strideynet/bsky-furry-feed

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>

Add commands for publish, unpublish, and tidying

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>

Some adjustment to the cursor and sorting

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>

Fix Docker release issues

Add the org/x/crypto/x509roots/fallback module to get statically linked
certs in the go binary. This is necessary as a Docker scratch image does
not have any system certs to fall back on.

Fix regression in the db migration scripts.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>

Add github action to run goreleaser

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>

Fix go version in github action

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>

Add publish workflow

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>

fix publish action

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>