build: Only build for 64 bit architectures
For some reason the modernc.org/sqlite library produces errors when
building for i382 and arm6 architectures. If we only build for amd64 and
arm64 architectures the goreleaser build job works.

Mostly everyone who runs a server environment will be running on 64 bit
architectures, so compiling for 32 bit is also wasteful. The build job
should take a lot less time when we reduce the number of binaries
significantly.