fix: Actually run delete SQL statement
The delete statement was previously never executed. Now the tidy
function executes the statement. The calculation of the time was also
off as it was checking against number of days in the future. We want to
check against days in the past so needed to add a negative multiplier to
get negative time.