fix: Continue from last processed seq add rollback cmd
When processing commits/events from the ATProto firehose we also get a
sequence number. This number can be used as a cursor into the feed to
continue from a last known point in case the feed server goes down, or
similar.

The feed cursor allows you to subscribe to the feed with backfill
(within reason) from the provided seq cursor.

Norsky stores the sequence number in the sequence table which is limited
to exactly one row. There is an id column to enforce the primary id, and
a placeholder value is inserted upon migration to ensure there always is
a value to update from.

We only update the sequence number if we processed a create post event
as otherwise there would be excessive amounts of write on the DB.

Commit also adds rollback command which is useful when testing or if you
want to run an earlier version of Norsky.