fix: Better context handling and graceful shutdown
Use go contexts to handle shutting down go routines and closing
channels. This largely makes the receiver pattern with the subscriber
and firehose structs unecessary. Instead we pass the necessary arguments
to the functions together with the context. The functions listen for the
context to signal the process should close, so no separate shutdown
function is necessary.

Fiber is handled as before.

The main.go main function now sets up a cancelable context and passes
this to the urfave cli run command. A go routine listens for interrupts
and if so gracefully signals to shutdown by calling cancel on the
context.