fix: Firehose liveness probe
Sometimes the firehose will go silent without any errors being returned
anywhere. To avoid the feed server simply stopping up the subscribe
function now accepts a ticker that is reset for every new post the
Firehose receives. If the ticker ever reaches its 5 minute mark a
separate go routine in the subcribe and serve command modules will reset
the firehose and make a new connection.

This way if the firehose is silent for 5 minutes or more the Norsky
server will automatically reconnect.