perf: Use autoscaling event processor from indigo
Instead of processing all events with the sequential scheduler we now
process them with the auto-scaling scheduler. This scheduler with
flexibly measure the throughput of each worker and add or remove workers
as needed. This should offer flexible upscaling as the firehose
throughput grows and shrinks.

The sequence counting might be out of order with parallell processing.
This is not a big issue as it should not be that big of a gap and
reprocessing some records in the event where the feed generator needs to
reconnect to the firehose does not pose any issues.