fix: Keyword based search
Use ts_vector @@ websearch_to_tsquery to support more flexible queries
with spaces, etc. Surround each term from the feed config an or to allow
multiple keywords in the search. Wrap with quotes to allow white spaces
in terms.

Fix the parent_uri column to ensure either null value or a non-empty
string. This way we only need to query for IS NULL or IS NOT NULL.