fix: Negative keyword text search bug
It seems the websearch_to_tsquery did not work quite as I expected.
Supplying the exclusion keywords as "AND NOT (foo* OR bar*)" did not
work. After testing in pgAdmin it seems we instead need to split the
inclusion and exclusion keywords into two separate web search queries in
the Postgres query. The second web search is wrapped in AND NOT (). This
way rows matching the exclusion web search are removed from the result
set.